refactor: `makeCard` 메서드를 Card 팩토리 메서드로 수정
