+++ app/src/main/java/woowacourse/payments/domain/Card.kt
@@ -5,4 +5,21 @@ class Card(
     val expiredDate: ExpiredDate,
     val ownerName: OwnerName,
     val password: Password,
-)
+) {
+    companion object {
+        fun from(
+            cardNumber: String,
+            expiredDate: String,
+            ownerName: String,
+            password: String,
+        ): Result<Card> =
+            runCatching {
+                val cardNumber = CardNumber(cardNumber)
+                val expiredDate = ExpiredDate.of(expiredDate) ?: throw IllegalArgumentException()
+                val ownerName = OwnerName(ownerName)
+                val password = Password(password)
+
+                Card(cardNumber, expiredDate, ownerName, password)
+            }
+    }
+}
